Move giraffe ui elements
