+++ src/giraffe-ui/nav/nav.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 
-import CloseSvg from '../../svgs/close.svg'
+import CloseSvg from '../svgs/close.svg'
 
 export const MoNav = ({ isOpenMobile, children, close }) => (
   <div
@@ -13,8 +13,6 @@ export const MoNav = ({ isOpenMobile, children, close }) => (
     {children}
     <button
       onClick={close}
-      // Can remove style when merged: https://github.com/MoveOnOrg/giraffe/pull/3
-      style={{ left: 'auto' }}
       className='mo-nav__close'
     >
       <CloseSvg />
